Move resume name/edit to sidebar
