+++ src/components/Toolbar.tsx
@@ -1,68 +1,21 @@
-import { CloudDownload, CodeSquare, Edit2, Check, X } from "lucide-react";
+import { CloudDownload, CodeSquare, } from "lucide-react";
 import StatusIndicator from "./StatusIndicator";
-import { useState, useRef, useEffect } from "react";
 
 interface ToolbarProps {
   error: string | null;
   isLoading?: boolean;
   pageRendered?: boolean;
-  resumeName: string;
   onDownloadPdf: () => void;
   onDownloadLaTeX: () => void;
-  onResumeNameChange: (name: string) => void;
 }
 
 function Toolbar({
   error,
   isLoading,
   pageRendered,
-  resumeName,
   onDownloadPdf,
-  onDownloadLaTeX,
-  onResumeNameChange
+  onDownloadLaTeX
 }: ToolbarProps) {
-  const [isEditingName, setIsEditingName] = useState(false);
-  const [tempName, setTempName] = useState(resumeName);
-  const inputRef = useRef<HTMLInputElement>(null);
-
-  useEffect(() => {
-    setTempName(resumeName);
-  }, [resumeName]);
-
-  useEffect(() => {
-    if (isEditingName && inputRef.current) {
-      inputRef.current.focus();
-      inputRef.current.select();
-    }
-  }, [isEditingName]);
-
-  const handleStartEdit = () => {
-    setTempName(resumeName);
-    setIsEditingName(true);
-  };
-
-  const handleSaveEdit = () => {
-    const trimmedName = tempName.trim();
-    if (trimmedName && trimmedName !== resumeName) {
-      onResumeNameChange(trimmedName);
-    }
-    setIsEditingName(false);
-  };
-
-  const handleCancelEdit = () => {
-    setTempName(resumeName);
-    setIsEditingName(false);
-  };
-
-  const handleKeyDown = (e: React.KeyboardEvent) => {
-    if (e.key === 'Enter') {
-      e.preventDefault();
-      handleSaveEdit();
-    } else if (e.key === 'Escape') {
-      e.preventDefault();
-      handleCancelEdit();
-    }
-  };
 
   const buttonCss =
     `transition-colors
@@ -86,52 +39,8 @@ function Toolbar({
 
         {/* Desktop layout - all in one row */}
         <div className="hidden lg:flex lg:flex-row lg:justify-between rounded-lg max-w-[800px] mx-auto px-4 md:px-0" role="group">
-          <div className="flex-1 inline-flex items-center space-x-3">
+          <div className="flex-1 inline-flex items-center">
             <StatusIndicator error={error} isLoading={isLoading} pageRendered={pageRendered} />
-            
-            {/* Resume name editor */}
-            <div className="flex items-center">
-              {isEditingName ? (
-                <div className="flex items-center space-x-1">
-                  <input
-                    ref={inputRef}
-                    type="text"
-                    value={tempName}
-                    onChange={(e) => setTempName(e.target.value)}
-                    onKeyDown={handleKeyDown}
-                    className="px-2 py-1 text-sm bg-white dark:bg-zinc-700 border border-gray-300 dark:border-zinc-600 rounded text-gray-900 dark:text-gray-100 focus:outline-none focus:ring-2 focus:ring-purple-500 min-w-[200px]"
-                    placeholder="Resume name"
-                  />
-                  <button
-                    onClick={handleSaveEdit}
-                    className="p-1 text-green-600 hover:text-green-700 dark:text-green-400 dark:hover:text-green-300"
-                    title="Save name"
-                  >
-                    <Check className="size-4" />
-                  </button>
-                  <button
-                    onClick={handleCancelEdit}
-                    className="p-1 text-red-600 hover:text-red-700 dark:text-red-400 dark:hover:text-red-300"
-                    title="Cancel edit"
-                  >
-                    <X className="size-4" />
-                  </button>
-                </div>
-              ) : (
-                <div className="flex items-center space-x-1 group">
-                  <span className="text-sm text-gray-700 dark:text-gray-300 font-medium">
-                    {resumeName}
-                  </span>
-                  <button
-                    onClick={handleStartEdit}
-                    className="p-1 text-gray-500 hover:text-gray-700 dark:text-gray-400 dark:hover:text-gray-200 opacity-0 group-hover:opacity-100 transition-opacity"
-                    title="Edit resume name"
-                  >
-                    <Edit2 className="size-3" />
-                  </button>
-                </div>
-              )}
-            </div>
           </div>
 
           <div className="flex-1 flex justify-end">
@@ -160,59 +69,15 @@ function Toolbar({
           </div>
         </div>
 
-        {/* Mobile layout - three rows */}
+        {/* Mobile layout - two rows */}
         <div className="flex flex-col lg:hidden rounded-lg px-4" role="group">
           {/* First row - status indicator only */}
-          <div className="inline-flex items-center justify-center mb-2">
+          <div className="inline-flex items-center justify-center mb-0">
             <StatusIndicator error={error} isLoading={isLoading} pageRendered={pageRendered} />
           </div>
 
-          {/* Second row - resume name editor */}
-          <div className="flex items-center justify-center mb-2">
-            {isEditingName ? (
-              <div className="flex items-center space-x-1">
-                <input
-                  ref={inputRef}
-                  type="text"
-                  value={tempName}
-                  onChange={(e) => setTempName(e.target.value)}
-                  onKeyDown={handleKeyDown}
-                  className="px-2 py-1 text-sm bg-white dark:bg-zinc-700 border border-gray-300 dark:border-zinc-600 rounded text-gray-900 dark:text-gray-100 focus:outline-none focus:ring-2 focus:ring-purple-500 min-w-[200px]"
-                  placeholder="Resume name"
-                />
-                <button
-                  onClick={handleSaveEdit}
-                  className="p-1 text-green-600 hover:text-green-700 dark:text-green-400 dark:hover:text-green-300"
-                  title="Save name"
-                >
-                  <Check className="size-4" />
-                </button>
-                <button
-                  onClick={handleCancelEdit}
-                  className="p-1 text-red-600 hover:text-red-700 dark:text-red-400 dark:hover:text-red-300"
-                  title="Cancel edit"
-                >
-                  <X className="size-4" />
-                </button>
-              </div>
-            ) : (
-              <div className="flex items-center space-x-1 group">
-                <span className="text-sm text-gray-700 dark:text-gray-300 font-medium">
-                  {resumeName}
-                </span>
-                <button
-                  onClick={handleStartEdit}
-                  className="p-1 text-gray-500 hover:text-gray-700 dark:text-gray-400 dark:hover:text-gray-200"
-                  title="Edit resume name"
-                >
-                  <Edit2 className="size-3" />
-                </button>
-              </div>
-            )}
-          </div>
-
-          {/* Third row - buttons */}
-          <div className="flex flex-row justify-center items-center">
+          {/* Second row - buttons right */}
+          <div className="flex flex-row justify-between items-center">
 
             <div className="inline-flex">
               <button type="button"
